refactor: split zarlcode subsystem files
